Add files via upload
Hi this is  create record form with is not complete yet, the parts that are not complete are adding Priest ID and Beravement Minister ID to funeral mass plan. also Cemetery  However it insert Deceased table info, Primary contact info and mass plan info. also it connect them together by inserting the forint keys  I hope it is good as we want. it just has one problem which when the page refreshed it submit the information again. I did change small thing to database to let me insert to Deceased table. I upload  the file for you. so you can sole the problem if would like. when I woke up I will work on it if the problem not solve. 
happy long weekend